WIP Use forms hook
Frozen because validation errors don't show - might be because we were not using <form> wrapper, so added it, and now next shows "Error: Hydration failed because the initial UI does not match what was rendered on the server." when loading the page. Since this refactoring is pretty much aesthetical, freezing.